skills/stock-assistant/requirements.txt
# Python 3.6+ 兼容
aiohttp>=3.8.0,<4.0.0
websockets>=9.0,<13.0
aiofiles>=0.8.0
python-dotenv>=0.19.0
